refactor: split TokenTreemap.tsx into focused sub-components (#93)
Break down 1074-line monolithic TokenTreemap into 10 focused modules:
- tokenTreemap/constants.ts: shared types and constants
- tokenTreemap/useTokenTreemapData.ts: data fetching and state hook
- tokenTreemap/TreemapCell.tsx: SVG cell renderer
- tokenTreemap/TreemapTooltip.tsx: tooltip component
- tokenTreemap/PromptFeed.tsx: real-time prompt feed
- tokenTreemap/CostAnalysisPanel.tsx: cost analysis display
- tokenTreemap/PromptDetailModal.tsx: prompt detail modal
- tokenTreemap/ScanProgress.tsx: scan progress overlay
- tokenTreemap/DetailPanel.tsx: category detail panel
- tokenTreemap/ContextLogsPanel.tsx: context logs section
- tokenTreemap/index.ts: barrel exports

TokenTreemap.tsx is now a thin orchestrator (~160 lines).
All 35 existing tests pass with no changes.
