+++ src/components/TokenTreemap.tsx
@@ -7,614 +7,62 @@
  * + Real-time prompt feed (breaking news style)
  */
 
-import { useState, useEffect, useCallback, useRef } from 'react';
 import { Treemap, ResponsiveContainer, Tooltip } from 'recharts';
 import './TokenTreemap.css';
-import type { PromptHistoryItem, PromptAnalysisResult, CacheUsageItem, ScanTokensResult, ContextLogs } from '../types';
-
-// Treemap data type
-type TreemapNode = {
-  name: string;
-  originalName?: string; // Original category name (CATEGORY_INFO key)
-  size?: number;
-  tokens?: number;
-  percentage?: number;
-  color?: string;
-  children?: TreemapNode[];
-  // Cache details
-  cacheDetails?: CacheUsageItem[];
-  claudeMdPreview?: string;
-  // Status info
-  status?: 'warning' | 'good' | 'info' | 'neutral';
-  statusBadge?: string;
-};
-
+import { TreemapCell } from './tokenTreemap/TreemapCell';
+import { TreemapTooltip } from './tokenTreemap/TreemapTooltip';
+import { PromptFeed } from './tokenTreemap/PromptFeed';
+import { CostAnalysisPanel } from './tokenTreemap/CostAnalysisPanel';
+import { PromptDetailModal } from './tokenTreemap/PromptDetailModal';
+import { ScanProgress } from './tokenTreemap/ScanProgress';
+import { DetailPanel } from './tokenTreemap/DetailPanel';
+import { ContextLogsPanel } from './tokenTreemap/ContextLogsPanel';
+import { useTokenTreemapData } from './tokenTreemap/useTokenTreemapData';
+import { MODEL_PRICING } from './tokenTreemap/constants';
+import type { TreemapCellProps, ModelId } from './tokenTreemap/constants';
 
 type TokenTreemapProps = {
   onBack: () => void;
 };
 
-// Color palette (webpack-bundle-analyzer style)
-const COLORS = {
-  claudeMd: '#e74c3c',      // Red - CLAUDE.md (largest)
-  cacheCreation: '#f39c12', // Orange - cache creation
-  cacheRead: '#27ae60',     // Green - cache read (efficient)
-  userInput: '#3498db',     // Blue - user input
-  output: '#9b59b6',        // Purple - AI output
-  section: '#1abc9c',       // Teal - CLAUDE.md sections
-};
-
-// Friendly info per category
-const CATEGORY_INFO: Record<string, {
-  friendlyName: string;
-  icon: string;
-  description: string;
-  whatIsIt: string;
-  status: 'warning' | 'good' | 'info' | 'neutral';
-}> = {
-  'CLAUDE.md': {
-    friendlyName: '📄 Config File',
-    icon: '📄',
-    description: 'AI instructions sent with every request',
-    whatIsIt: 'CLAUDE.md is a set of instructions for the AI. It includes code style, rules, and conventions, and the entire content is sent with every request, incurring costs.',
-    status: 'warning',
-  },
-  'Cache Creation': {
-    friendlyName: '🆕 New Cache',
-    icon: '🆕',
-    description: 'Cache to be reused in subsequent requests',
-    whatIsIt: 'New context is being stored as cache. It costs now, but will be reused at a 90% discount from the next request onward.',
-    status: 'info',
-  },
-  'Cache Read': {
-    friendlyName: '♻️ Cached',
-    icon: '♻️',
-    description: '90% cost savings!',
-    whatIsIt: 'Content from previous conversations is being reused. Claude recognizes previously processed content and handles it quickly at a 90% discount.',
-    status: 'good',
-  },
-  'User Input': {
-    friendlyName: '💬 My Input',
-    icon: '💬',
-    description: 'Your actual input',
-    whatIsIt: 'Your actual prompts and conversation content. This is hard to reduce, but being concise and focused helps efficiency.',
-    status: 'neutral',
-  },
-  'AI Output': {
-    friendlyName: '🤖 AI Response',
-    icon: '🤖',
-    description: 'AI-generated response',
-    whatIsIt: 'Tokens generated by the AI. Output tokens cost 5x more than input, so request only what you need.',
-    status: 'neutral',
-  },
-};
-
-// Pricing per model (per 1M tokens)
-const MODEL_PRICING = {
-  'claude-sonnet-4-20250514': { name: 'Claude Sonnet 4', input: 3.0, output: 15.0, cacheRead: 0.3 },
-  'claude-3-5-sonnet-20241022': { name: 'Claude 3.5 Sonnet', input: 3.0, output: 15.0, cacheRead: 0.3 },
-  'claude-3-opus-20240229': { name: 'Claude 3 Opus', input: 15.0, output: 75.0, cacheRead: 1.5 },
-  'claude-3-haiku-20240307': { name: 'Claude 3 Haiku', input: 0.25, output: 1.25, cacheRead: 0.03 },
-  'claude-opus-4-20250514': { name: 'Claude Opus 4', input: 15.0, output: 75.0, cacheRead: 1.5 },
-} as const;
-
-type ModelId = keyof typeof MODEL_PRICING;
-
-// Custom Treemap cell renderer
-type TreemapCellProps = {
-  x: number;
-  y: number;
-  width: number;
-  height: number;
-  name?: string;
-  tokens?: number;
-  percentage?: number;
-  color?: string;
-  depth?: number;
-};
-
-const CustomTreemapContent = (props: TreemapCellProps) => {
-  const { x, y, width, height, name, tokens, percentage, color, depth } = props;
-
-  // Omit text for cells too small
-  const showText = width > 60 && height > 40;
-  const showDetails = width > 100 && height > 60;
-
-  return (
-    <g>
-      <rect
-        x={x}
-        y={y}
-        width={width}
-        height={height}
-        style={{
-          fill: color || '#8884d8',
-          stroke: '#1a1a2e',
-          strokeWidth: depth === 1 ? 3 : 1,
-          opacity: 0.9,
-          cursor: 'pointer',
-        }}
-      />
-      {showText && (
-        <>
-          <text
-            x={x + width / 2}
-            y={y + height / 2 - (showDetails ? 10 : 0)}
-            textAnchor="middle"
-            dominantBaseline="middle"
-            style={{
-              fontSize: Math.min(14, width / 8),
-              fill: '#fff',
-              fontWeight: 'bold',
-              textShadow: '1px 1px 2px rgba(0,0,0,0.8)',
-              pointerEvents: 'none',
-            }}
-          >
-            {name}
-          </text>
-          {showDetails && (
-            <>
-              <text
-                x={x + width / 2}
-                y={y + height / 2 + 8}
-                textAnchor="middle"
-                dominantBaseline="middle"
-                style={{
-                  fontSize: Math.min(11, width / 10),
-                  fill: '#ddd',
-                  pointerEvents: 'none',
-                }}
-              >
-                {tokens?.toLocaleString()} tokens
-              </text>
-              <text
-                x={x + width / 2}
-                y={y + height / 2 + 22}
-                textAnchor="middle"
-                dominantBaseline="middle"
-                style={{
-                  fontSize: Math.min(10, width / 12),
-                  fill: '#aaa',
-                  pointerEvents: 'none',
-                }}
-              >
-                {percentage?.toFixed(1)}%
-              </text>
-            </>
-          )}
-        </>
-      )}
-    </g>
-  );
-};
-
-// Custom tooltip (with cache details)
-type TreemapTooltipProps = {
-  active?: boolean;
-  payload?: Array<{ payload: TreemapNode }>;
-};
-
-const CustomTooltip = ({ active, payload }: TreemapTooltipProps) => {
-  if (!active || !payload || !payload.length) return null;
-
-  const data = payload[0].payload;
-
-  return (
-    <div className="treemap-tooltip">
-      <div className="tooltip-header">{data.name}</div>
-      <div className="tooltip-row">
-        <span className="label">Tokens:</span>
-        <span className="value">{data.tokens?.toLocaleString()}</span>
-      </div>
-      <div className="tooltip-row">
-        <span className="label">Percentage:</span>
-        <span className="value">{data.percentage?.toFixed(2)}%</span>
-      </div>
-      {data.name === 'CLAUDE.md' && (
-        <div className="tooltip-warning">
-          ⚠️ Takes up most of the context
-        </div>
-      )}
-      {data.name === 'Cache Read' && (
-        <div className="tooltip-good">
-          ✅ 90% cost savings from cache!
-        </div>
-      )}
-      {/* Cache details */}
-      {data.cacheDetails && data.cacheDetails.length > 0 && (
-        <div className="tooltip-cache-details">
-          <div className="cache-details-header">📋 Recent cache usage:</div>
-          {data.cacheDetails.slice(0, 3).map((item: CacheUsageItem, idx: number) => (
-            <div key={idx} className="cache-detail-item">
-              <span className="cache-prompt">{item.prompt || '(request)'}</span>
-              <span className="cache-tokens">
-                {item.cacheRead > 0 && `Read: ${item.cacheRead.toLocaleString()}`}
-                {item.cacheCreation > 0 && ` Created: ${item.cacheCreation.toLocaleString()}`}
-              </span>
-            </div>
-          ))}
-        </div>
-      )}
-      {/* CLAUDE.md preview */}
-      {data.claudeMdPreview && (
-        <div className="tooltip-claudemd-preview">
-          <div className="claudemd-preview-header">📄 Cached CLAUDE.md:</div>
-          <div className="claudemd-preview-content">{data.claudeMdPreview}</div>
-        </div>
-      )}
-    </div>
-  );
-};
-
 export const TokenTreemap = ({ onBack }: TokenTreemapProps) => {
-  const [isScanning, setIsScanning] = useState(false);
-  const [scanProgress, setScanProgress] = useState(0);
-  const [treemapData, setTreemapData] = useState<TreemapNode[]>([]);
-  const [totalTokens, setTotalTokens] = useState(0);
-  const [selectedNode, setSelectedNode] = useState<string | null>(null);
-  const [selectedModel, setSelectedModel] = useState<ModelId>('claude-sonnet-4-20250514');
-  const [_cacheInfo, setCacheInfo] = useState<ScanTokensResult['cacheInfo'] | null>(null);
-
-  // Prompt history state
-  const [promptHistory, setPromptHistory] = useState<PromptHistoryItem[]>([]);
-  const [selectedPrompt, setSelectedPrompt] = useState<string | null>(null);
-  const [promptAnalysis, setPromptAnalysis] = useState<PromptAnalysisResult | null>(null);
-  const [isAnalyzing, setIsAnalyzing] = useState(false);
-  const [showAllPrompts, setShowAllPrompts] = useState(false);
-  const [currentPage, setCurrentPage] = useState(0);
-  const [showPromptModal, setShowPromptModal] = useState(false);
-  const [modalPrompt, setModalPrompt] = useState<PromptHistoryItem | null>(null);
-  const PROMPTS_PER_PAGE = 20;
-  const VISIBLE_PROMPTS = 3;
-  const pollingRef = useRef<NodeJS.Timeout | null>(null);
-  const treemapPollingRef = useRef<NodeJS.Timeout | null>(null);
-
-  // Context logs state (referenced file list)
-  const [contextLogs, setContextLogs] = useState<ContextLogs | null>(null);
-  const [showContextLogs, setShowContextLogs] = useState(false);
-
-  // Detail panel state (on area click)
-  const [detailPanel, setDetailPanel] = useState<{
-    isOpen: boolean;
-    category: string;
-    tokens: number;
-    percentage: number;
-    cost: number;
-    savedCost?: number;
-  } | null>(null);
-
-  // Prompt click -> open modal
-  const handlePromptClick = useCallback((prompt: PromptHistoryItem) => {
-    setModalPrompt(prompt);
-    setShowPromptModal(true);
-  }, []);
-
-  // Start analysis from modal
-  const handleAnalyzeFromModal = useCallback(() => {
-    if (modalPrompt) {
-      setShowPromptModal(false);
-      analyzePrompt(modalPrompt.id);
-    }
-  }, [modalPrompt]);
-
-  // Prompt analysis and Treemap update
-  const analyzePrompt = useCallback(async (promptId: string) => {
-    setIsAnalyzing(true);
-    setSelectedPrompt(promptId);
-    try {
-      const analysis = await window.api.analyzePrompt?.(promptId);
-      setPromptAnalysis(analysis);
-
-      // Update Treemap data
-      if (analysis?.response) {
-        const { response } = analysis;
-        const total = response.inputTokens + response.outputTokens + response.cacheCreationTokens;
-
-        const newTreemapData: TreemapNode[] = [
-          {
-            name: 'Input Tokens',
-            size: response.inputTokens,
-            tokens: response.inputTokens,
-            percentage: (response.inputTokens / total) * 100,
-            color: COLORS.userInput,
-          },
-          {
-            name: 'Output Tokens',
-            size: response.outputTokens,
-            tokens: response.outputTokens,
-            percentage: (response.outputTokens / total) * 100,
-            color: COLORS.output,
-          },
-          {
-            name: 'Cache Creation',
-            size: response.cacheCreationTokens,
-            tokens: response.cacheCreationTokens,
-            percentage: (response.cacheCreationTokens / total) * 100,
-            color: COLORS.cacheCreation,
-          },
-          {
-            name: 'Cache Read',
-            size: response.cacheReadTokens,
-            tokens: response.cacheReadTokens,
-            percentage: (response.cacheReadTokens / (total + response.cacheReadTokens)) * 100,
-            color: COLORS.cacheRead,
-          },
-        ].filter(node => (node.size || 0) > 0);
-
-        setTreemapData(newTreemapData);
-        setTotalTokens(total);
-
-      }
-    } catch (error) {
-      console.error('Analyze error:', error);
-      setPromptAnalysis(null);
-    } finally {
-      setIsAnalyzing(false);
-    }
-  }, [selectedModel]);
-
-  // Start scan
-  const startScan = useCallback(async () => {
-    setIsScanning(true);
-    setScanProgress(0);
-
-    try {
-      // Progress animation
-      for (let i = 0; i <= 100; i += 5) {
-        setScanProgress(i);
-        await new Promise(r => setTimeout(r, 30));
-      }
-
-      // API call
-      const data: ScanTokensResult = await window.api.scanTokens();
-
-      if (data?.breakdown) {
-        const { breakdown } = data;
-        setTotalTokens(breakdown.total);
-
-        // Save cache info
-        if (data.cacheInfo) {
-          setCacheInfo(data.cacheInfo);
-        }
-
-        // Build Treemap data (friendly names + status included)
-        const claudeMdPercentage = (breakdown.claudeMd.total / breakdown.total) * 100;
-        const treeData = ([
-          {
-            name: CATEGORY_INFO['CLAUDE.md'].friendlyName,
-            originalName: 'CLAUDE.md',
-            size: breakdown.claudeMd.total,
-            tokens: breakdown.claudeMd.total,
-            percentage: claudeMdPercentage,
-            color: COLORS.claudeMd,
-            claudeMdPreview: data.cacheInfo?.claudeMdPreview?.slice(0, 200),
-            status: (claudeMdPercentage > 40 ? 'warning' : 'neutral') as TreemapNode['status'],
-            statusBadge: claudeMdPercentage > 40 ? '⚠️ Top cost driver' : undefined,
-            children: data.claudeMdSections?.map((section, idx) => ({
-              name: section.section,
-              size: section.tokens,
-              tokens: section.tokens,
-              percentage: section.percentage,
-              color: `hsl(${(idx * 30) % 360}, 70%, 50%)`,
-            })),
-          },
-          {
-            name: CATEGORY_INFO['Cache Creation'].friendlyName,
-            originalName: 'Cache Creation',
-            size: breakdown.cacheCreation,
-            tokens: breakdown.cacheCreation,
-            percentage: (breakdown.cacheCreation / breakdown.total) * 100,
-            color: COLORS.cacheCreation,
-            status: 'info',
-            statusBadge: 'Reused next time',
-            cacheDetails: data.cacheInfo?.recentCacheUsage?.filter(c => c.cacheCreation > 0),
-          },
-          {
-            name: CATEGORY_INFO['Cache Read'].friendlyName,
-            originalName: 'Cache Read',
-            size: breakdown.cacheRead,
-            tokens: breakdown.cacheRead,
-            percentage: (breakdown.cacheRead / breakdown.total) * 100,
-            color: COLORS.cacheRead,
-            status: 'good',
-            statusBadge: '✅ 90% savings!',
-            cacheDetails: data.cacheInfo?.recentCacheUsage?.filter(c => c.cacheRead > 0),
-          },
-          {
-            name: CATEGORY_INFO['User Input'].friendlyName,
-            originalName: 'User Input',
-            size: breakdown.userInput,
-            tokens: breakdown.userInput,
-            percentage: (breakdown.userInput / breakdown.total) * 100,
-            color: COLORS.userInput,
-            status: 'neutral',
-          },
-          {
-            name: CATEGORY_INFO['AI Output'].friendlyName,
-            originalName: 'AI Output',
-            size: breakdown.output,
-            tokens: breakdown.output,
-            percentage: (breakdown.output / breakdown.total) * 100,
-            color: COLORS.output,
-            status: 'neutral',
-          },
-        ] as TreemapNode[]).filter(node => (node.size || 0) > 0);
-
-        setTreemapData(treeData);
-      }
-    } catch (error) {
-      console.error('Scan error:', error);
-      // Demo data (with friendly names)
-      setTreemapData([
-        { name: '📄 Config File', originalName: 'CLAUDE.md', size: 15234, tokens: 15234, percentage: 62.3, color: COLORS.claudeMd, status: 'warning', statusBadge: '⚠️ Top cost driver' },
-        { name: '🆕 New Cache', originalName: 'Cache Creation', size: 5234, tokens: 5234, percentage: 21.4, color: COLORS.cacheCreation, status: 'info', statusBadge: 'Reused next time' },
-        { name: '♻️ Cached', originalName: 'Cache Read', size: 2891, tokens: 2891, percentage: 11.8, color: COLORS.cacheRead, status: 'good', statusBadge: '✅ 90% savings!' },
-        { name: '💬 My Input', originalName: 'User Input', size: 512, tokens: 512, percentage: 2.1, color: COLORS.userInput, status: 'neutral' },
-        { name: '🤖 AI Response', originalName: 'AI Output', size: 589, tokens: 589, percentage: 2.4, color: COLORS.output, status: 'neutral' },
-      ]);
-      setTotalTokens(24460);
-    } finally {
-      setIsScanning(false);
-    }
-  }, []);
-
-  // Fetch context logs (referenced file list)
-  const fetchContextLogs = useCallback(async () => {
-    try {
-      const logs = await window.api.getContextLogs?.();
-      if (logs) {
-        setContextLogs(logs);
-      }
-    } catch (error) {
-      console.error('Get context logs error:', error);
-    }
-  }, []);
-
-  // Fetch prompt history
-  const fetchPromptHistory = useCallback(async () => {
-    try {
-      const history = await window.api.getPromptHistory?.();
-      if (history && Array.isArray(history)) {
-        setPromptHistory(history);
-      }
-    } catch (error) {
-      // Demo data
-      const demoHistory: PromptHistoryItem[] = [
-        {
-          id: '1',
-          timestamp: new Date().toISOString(),
-          content: 'What about showing your requests at the top like a live chat? Like breaking news, with 3 lines...',
-          tokens: 1234,
-        },
-        {
-          id: '2',
-          timestamp: new Date(Date.now() - 60000).toISOString(),
-          content: 'How about something like webpack-bundle-analyzer?',
-          tokens: 892,
-        },
-        {
-          id: '3',
-          timestamp: new Date(Date.now() - 120000).toISOString(),
-          content: 'Oh yeah, make it Treemap style',
-          tokens: 456,
-        },
-      ];
-      setPromptHistory(demoHistory);
-    }
-  }, []);
-
-  // Real-time polling
-  useEffect(() => {
-    fetchPromptHistory();
-    fetchContextLogs();
-
-    // Check for new prompts every 1 second (real-time)
-    pollingRef.current = setInterval(() => {
-      fetchPromptHistory();
-      fetchContextLogs();
-    }, 1000);
-
-    return () => {
-      if (pollingRef.current) {
-        clearInterval(pollingRef.current);
-      }
-    };
-  }, [fetchPromptHistory, fetchContextLogs]);
-
-  // Silently update Treemap data (without animation)
-  const silentScan = useCallback(async () => {
-    // Skip if analyzing a prompt (do not overwrite while user is analyzing)
-    if (selectedPrompt || isAnalyzing) return;
-
-    try {
-      const data: ScanTokensResult = await window.api.scanTokens();
-
-      if (data?.breakdown) {
-        const { breakdown } = data;
-        setTotalTokens(breakdown.total);
-
-        // Save cache info
-        if (data.cacheInfo) {
-          setCacheInfo(data.cacheInfo);
-        }
-
-        const treeData: TreemapNode[] = [
-          {
-            name: 'CLAUDE.md',
-            size: breakdown.claudeMd.total,
-            tokens: breakdown.claudeMd.total,
-            percentage: (breakdown.claudeMd.total / breakdown.total) * 100,
-            color: COLORS.claudeMd,
-            claudeMdPreview: data.cacheInfo?.claudeMdPreview?.slice(0, 200),
-            children: data.claudeMdSections?.map((section, idx) => ({
-              name: section.section,
-              size: section.tokens,
-              tokens: section.tokens,
-              percentage: section.percentage,
-              color: `hsl(${(idx * 30) % 360}, 70%, 50%)`,
-            })),
-          },
-          {
-            name: 'Cache Creation',
-            size: breakdown.cacheCreation,
-            tokens: breakdown.cacheCreation,
-            percentage: (breakdown.cacheCreation / breakdown.total) * 100,
-            color: COLORS.cacheCreation,
-            cacheDetails: data.cacheInfo?.recentCacheUsage?.filter(c => c.cacheCreation > 0),
-          },
-          {
-            name: 'Cache Read',
-            size: breakdown.cacheRead,
-            tokens: breakdown.cacheRead,
-            percentage: (breakdown.cacheRead / breakdown.total) * 100,
-            color: COLORS.cacheRead,
-            cacheDetails: data.cacheInfo?.recentCacheUsage?.filter(c => c.cacheRead > 0),
-          },
-          {
-            name: 'User Input',
-            size: breakdown.userInput,
-            tokens: breakdown.userInput,
-            percentage: (breakdown.userInput / breakdown.total) * 100,
-            color: COLORS.userInput,
-          },
-          {
-            name: 'AI Output',
-            size: breakdown.output,
-            tokens: breakdown.output,
-            percentage: (breakdown.output / breakdown.total) * 100,
-            color: COLORS.output,
-          },
-        ].filter(node => (node.size || 0) > 0);
-
-        setTreemapData(treeData);
-      }
-    } catch (error) {
-      console.error('Silent scan error:', error);
-    }
-  }, [selectedPrompt, isAnalyzing]);
-
-  // Auto-scan on component mount + periodic updates
-  useEffect(() => {
-    startScan();
-
-    // Update Treemap data every 5 seconds
-    treemapPollingRef.current = setInterval(() => {
-      silentScan();
-    }, 5000);
-
-    return () => {
-      if (treemapPollingRef.current) {
-        clearInterval(treemapPollingRef.current);
-      }
-    };
-  }, [silentScan]);
+  const {
+    isScanning,
+    scanProgress,
+    treemapData,
+    totalTokens,
+    selectedNode,
+    selectedModel,
+    setSelectedModel,
+    startScan,
+    promptHistory,
+    selectedPrompt,
+    promptAnalysis,
+    isAnalyzing,
+    showAllPrompts,
+    currentPage,
+    setCurrentPage,
+    showPromptModal,
+    modalPrompt,
+    contextLogs,
+    showContextLogs,
+    detailPanel,
+    handlePromptClick,
+    handleAnalyzeFromModal,
+    handleCloseAnalysis,
+    handleCloseModal,
+    handleToggleContextLogs,
+    handleCloseDetailPanel,
+    handleToggleShowAllPrompts,
+    handleLegendClick,
+  } = useTokenTreemapData();
 
   return (
     <div className="token-treemap">
       {/* Header */}
       <div className="treemap-header">
         <button className="back-btn" onClick={onBack}>← Back</button>
-        <h2>📦 Token Analyzer</h2>
+        <h2>Token Analyzer</h2>
         <div className="header-actions">
           <select
             className="model-select"
@@ -637,217 +85,44 @@ export const TokenTreemap = ({ onBack }: TokenTreemapProps) => {
         </div>
       </div>
 
-      {/* Real-time prompt feed (breaking news style) */}
-      <div className="prompt-feed">
-        <div className="feed-header">
-          <span className="feed-title">📡 Live Prompts</span>
-          <button
-            className="feed-toggle"
-            onClick={() => setShowAllPrompts(!showAllPrompts)}
-          >
-            {showAllPrompts ? 'Collapse' : `View all (${promptHistory.length})`}
-          </button>
-        </div>
+      {/* Real-time prompt feed */}
+      <PromptFeed
+        promptHistory={promptHistory}
+        selectedPrompt={selectedPrompt}
+        showAllPrompts={showAllPrompts}
+        currentPage={currentPage}
+        onToggleShowAll={handleToggleShowAllPrompts}
+        onPageChange={setCurrentPage}
+        onPromptClick={handlePromptClick}
+      />
 
-        <div className="feed-list">
-          {(showAllPrompts
-            ? promptHistory.slice(currentPage * PROMPTS_PER_PAGE, (currentPage + 1) * PROMPTS_PER_PAGE)
-            : promptHistory.slice(0, VISIBLE_PROMPTS)
-          ).map((prompt) => (
-            <div
-              key={prompt.id}
-              className={`feed-item ${selectedPrompt === prompt.id ? 'selected' : ''}`}
-              onClick={() => handlePromptClick(prompt)}
-            >
-              <span className="feed-time">
-                {new Date(prompt.timestamp).toLocaleTimeString('en-US', {
-                  hour: '2-digit',
-                  minute: '2-digit',
-                })}
-              </span>
-              <span className="feed-content">{prompt.content}</span>
-              <span className="feed-tokens">{prompt.tokens.toLocaleString()} tok</span>
-            </div>
-          ))}
-        </div>
-
-        {/* Pagination */}
-        {showAllPrompts && promptHistory.length > PROMPTS_PER_PAGE && (
-          <div className="feed-pagination">
-            <button
-              disabled={currentPage === 0}
-              onClick={() => setCurrentPage(p => p - 1)}
-            >
-              ← Prev
-            </button>
-            <span>
-              {currentPage + 1} / {Math.ceil(promptHistory.length / PROMPTS_PER_PAGE)}
-            </span>
-            <button
-              disabled={(currentPage + 1) * PROMPTS_PER_PAGE >= promptHistory.length}
-              onClick={() => setCurrentPage(p => p + 1)}
-            >
-              Next →
-            </button>
-          </div>
-        )}
-      </div>
-
-      {/* Prompt analysis results */}
+      {/* Cost analysis results */}
       {promptAnalysis && !isAnalyzing && (
-        <div className="prompt-analysis">
-          <div className="analysis-header">
-            <span className="analysis-title">💰 Cost Analysis</span>
-            <button
-              className="analysis-close"
-              onClick={() => {
-                setPromptAnalysis(null);
-                setSelectedPrompt(null);
-              }}
-            >
-              ✕
-            </button>
-          </div>
-
-          <div className="analysis-prompt">
-            <div className="analysis-label">Prompt</div>
-            <div className="analysis-content">
-              {promptAnalysis.prompt.content}
-            </div>
-            <div className="analysis-meta">
-              {promptAnalysis.prompt.tokens.toLocaleString()} tokens
-            </div>
-          </div>
-
-          {promptAnalysis.response && (
-            <div className="analysis-response">
-              <div className="analysis-label">Response ({promptAnalysis.response.model})</div>
-              <div className="analysis-tokens">
-                <div className="token-item">
-                  <span className="token-label">Input</span>
-                  <span className="token-value">{promptAnalysis.response.inputTokens.toLocaleString()}</span>
-                </div>
-                <div className="token-item">
-                  <span className="token-label">Output</span>
-                  <span className="token-value">{promptAnalysis.response.outputTokens.toLocaleString()}</span>
-                </div>
-                <div className="token-item cache">
-                  <span className="token-label">Cache Read</span>
-                  <span className="token-value">{promptAnalysis.response.cacheReadTokens.toLocaleString()}</span>
-                </div>
-                <div className="token-item">
-                  <span className="token-label">Cache Creation</span>
-                  <span className="token-value">{promptAnalysis.response.cacheCreationTokens.toLocaleString()}</span>
-                </div>
-              </div>
-            </div>
-          )}
-
-          <div className="analysis-cost">
-            {(() => {
-              const pricing = MODEL_PRICING[selectedModel];
-              const resp = promptAnalysis.response;
-              if (!resp) return null;
-
-              const inputCost = (resp.inputTokens / 1000000) * pricing.input;
-              const outputCost = (resp.outputTokens / 1000000) * pricing.output;
-              const cacheCost = (resp.cacheReadTokens / 1000000) * pricing.cacheRead;
-              const totalCost = inputCost + outputCost + cacheCost;
-              const savedCost = ((resp.cacheReadTokens / 1000000) * pricing.input) - cacheCost;
-
-              return (
-                <>
-                  <div className="cost-model">
-                    Based on {pricing.name}
-                  </div>
-                  <div className="cost-row">
-                    <span>Input Cost</span>
-                    <span>${inputCost.toFixed(6)}</span>
-                  </div>
-                  <div className="cost-row">
-                    <span>Output Cost</span>
-                    <span>${outputCost.toFixed(6)}</span>
-                  </div>
-                  <div className="cost-row cache">
-                    <span>Cache Cost</span>
-                    <span>${cacheCost.toFixed(6)}</span>
-                  </div>
-                  <div className="cost-row total">
-                    <span>Total Cost</span>
-                    <span>${totalCost.toFixed(6)}</span>
-                  </div>
-                  {savedCost > 0 && (
-                    <div className="cost-row saved">
-                      <span>💰 Cache Savings</span>
-                      <span>-${savedCost.toFixed(6)}</span>
-                    </div>
-                  )}
-                </>
-              );
-            })()}
-          </div>
-        </div>
+        <CostAnalysisPanel
+          promptAnalysis={promptAnalysis}
+          selectedModel={selectedModel}
+          onClose={handleCloseAnalysis}
+        />
       )}
 
       {isAnalyzing && (
         <div className="analyzing-indicator">
-          <span className="analyzing-spinner">⏳</span>
+          <span className="analyzing-spinner">...</span>
           Analyzing...
         </div>
       )}
 
       {/* Prompt detail modal */}
       {showPromptModal && modalPrompt && (
-        <div className="prompt-modal-overlay" onClick={() => setShowPromptModal(false)}>
-          <div className="prompt-modal" onClick={(e) => e.stopPropagation()}>
-            <div className="modal-header">
-              <span className="modal-title">📝 Prompt Details</span>
-              <button className="modal-close" onClick={() => setShowPromptModal(false)}>✕</button>
-            </div>
-            <div className="modal-meta">
-              <span className="modal-time">
-                {new Date(modalPrompt.timestamp).toLocaleString('en-US')}
-              </span>
-              <span className="modal-tokens">{modalPrompt.tokens.toLocaleString()} tokens</span>
-            </div>
-            <div className="modal-content">
-              {modalPrompt.fullContent || modalPrompt.content}
-            </div>
-            <div className="modal-actions">
-              <button className="modal-btn secondary" onClick={() => setShowPromptModal(false)}>
-                Close
-              </button>
-              <button className="modal-btn primary" onClick={handleAnalyzeFromModal}>
-                💰 Cost Analysis
-              </button>
-            </div>
-          </div>
-        </div>
+        <PromptDetailModal
+          prompt={modalPrompt}
+          onClose={handleCloseModal}
+          onAnalyze={handleAnalyzeFromModal}
+        />
       )}
 
       {/* Scan progress */}
-      {isScanning && (
-        <div className="scan-overlay">
-          <div className="scan-modal">
-            <div className="scan-animation">
-              <div className="scan-line" style={{ top: `${scanProgress}%` }} />
-              <div className="scan-grid">
-                {[...Array(20)].map((_, i) => (
-                  <div
-                    key={i}
-                    className="scan-cell"
-                    style={{
-                      opacity: scanProgress > (i * 5) ? 1 : 0.2,
-                      backgroundColor: `hsl(${i * 18}, 70%, 50%)`,
-                    }}
-                  />
-                ))}
-              </div>
-            </div>
-            <div className="scan-text">Analyzing tokens... {scanProgress}%</div>
-          </div>
-        </div>
-      )}
+      {isScanning && <ScanProgress scanProgress={scanProgress} />}
 
       {/* Treemap */}
       {!isScanning && treemapData.length > 0 && (
@@ -859,56 +134,36 @@ export const TokenTreemap = ({ onBack }: TokenTreemapProps) => {
                 dataKey="size"
                 aspectRatio={4 / 3}
                 stroke="#1a1a2e"
-                content={((props: Record<string, unknown>) => <CustomTreemapContent {...(props as TreemapCellProps)} />) as unknown as React.ReactElement}
+                content={((props: Record<string, unknown>) => <TreemapCell {...(props as TreemapCellProps)} />) as unknown as React.ReactElement}
               >
-                <Tooltip content={<CustomTooltip />} />
+                <Tooltip content={<TreemapTooltip />} />
               </Treemap>
             </ResponsiveContainer>
           </div>
 
-          {/* Legend (click for detail panel) */}
+          {/* Legend */}
           <div className="treemap-legend">
-            {treemapData.map((node) => {
-              const pricing = MODEL_PRICING[selectedModel];
-              const cost = ((node.tokens || 0) / 1000000) * pricing.input;
-              const savedCost = node.originalName === 'Cache Read'
-                ? ((node.tokens || 0) / 1000000) * (pricing.input - pricing.cacheRead)
-                : 0;
-
-              return (
-                <div
-                  key={node.name}
-                  className={`legend-item ${selectedNode === node.name ? 'selected' : ''} status-${node.status || 'neutral'}`}
-                  onClick={() => {
-                    setSelectedNode(selectedNode === node.name ? null : node.name);
-                    if (node.originalName && CATEGORY_INFO[node.originalName]) {
-                      setDetailPanel({
-                        isOpen: true,
-                        category: node.originalName,
-                        tokens: node.tokens || 0,
-                        percentage: node.percentage || 0,
-                        cost,
-                        savedCost: savedCost > 0 ? savedCost : undefined,
-                      });
-                    }
-                  }}
-                >
-                  <span
-                    className="legend-color"
-                    style={{ backgroundColor: node.color }}
-                  />
-                  <span className="legend-name">{node.name}</span>
-                  {node.statusBadge && (
-                    <span className={`legend-badge status-${node.status}`}>
-                      {node.statusBadge}
-                    </span>
-                  )}
-                  <span className="legend-value">
-                    {node.tokens?.toLocaleString()} ({node.percentage?.toFixed(1)}%)
+            {treemapData.map((node) => (
+              <div
+                key={node.name}
+                className={`legend-item ${selectedNode === node.name ? 'selected' : ''} status-${node.status || 'neutral'}`}
+                onClick={() => handleLegendClick(node)}
+              >
+                <span
+                  className="legend-color"
+                  style={{ backgroundColor: node.color }}
+                />
+                <span className="legend-name">{node.name}</span>
+                {node.statusBadge && (
+                  <span className={`legend-badge status-${node.status}`}>
+                    {node.statusBadge}
                   </span>
-                </div>
-              );
-            })}
+                )}
+                <span className="legend-value">
+                  {node.tokens?.toLocaleString()} ({node.percentage?.toFixed(1)}%)
+                </span>
+              </div>
+            ))}
           </div>
 
           {/* Total */}
@@ -916,158 +171,24 @@ export const TokenTreemap = ({ onBack }: TokenTreemapProps) => {
             <span className="total-label">Total Tokens:</span>
             <span className="total-value">{totalTokens.toLocaleString()}</span>
           </div>
-
         </>
       )}
 
-      {/* Detail panel (on area click) */}
-      {detailPanel?.isOpen && detailPanel.category && CATEGORY_INFO[detailPanel.category] && (
-        <div className="detail-panel-overlay" onClick={() => setDetailPanel(null)}>
-          <div className="detail-panel" onClick={(e) => e.stopPropagation()}>
-            <div className="detail-panel-header">
-              <h3>{CATEGORY_INFO[detailPanel.category].friendlyName}</h3>
-              <button className="detail-panel-close" onClick={() => setDetailPanel(null)}>✕</button>
-            </div>
-
-            <div className="detail-panel-stats">
-              <div className="stat-item">
-                <span className="stat-label">Tokens</span>
-                <span className="stat-value">{detailPanel.tokens.toLocaleString()}</span>
-              </div>
-              <div className="stat-item">
-                <span className="stat-label">Share</span>
-                <span className="stat-value">{detailPanel.percentage.toFixed(1)}%</span>
-              </div>
-              <div className="stat-item">
-                <span className="stat-label">Cost</span>
-                <span className="stat-value">${detailPanel.cost.toFixed(4)}</span>
-              </div>
-              {detailPanel.savedCost && (
-                <div className="stat-item good">
-                  <span className="stat-label">Saved</span>
-                  <span className="stat-value">-${detailPanel.savedCost.toFixed(4)}</span>
-                </div>
-              )}
-            </div>
-
-            <div className="detail-panel-section">
-              <h4>❓ What is this?</h4>
-              <p>{CATEGORY_INFO[detailPanel.category].whatIsIt}</p>
-            </div>
-
-            {detailPanel.category === 'CLAUDE.md' && (
-              <div className="detail-panel-actions">
-                <button
-                  className="detail-action-btn"
-                  onClick={() => {
-                    // Open CLAUDE.md file (future implementation)
-                    alert('CLAUDE.md editing will be available in a future update.');
-                  }}
-                >
-                  📝 Open CLAUDE.md
-                </button>
-              </div>
-            )}
-          </div>
-        </div>
+      {/* Detail panel */}
+      {detailPanel?.isOpen && (
+        <DetailPanel
+          data={detailPanel}
+          onClose={handleCloseDetailPanel}
+        />
       )}
 
-      {/* Referenced context - always displayed independently of Treemap */}
+      {/* Context logs */}
       {contextLogs && !isScanning && (
-        <div className="context-logs-section" style={{ marginTop: '20px' }}>
-          <div
-            className="context-logs-header"
-            onClick={() => setShowContextLogs(!showContextLogs)}
-          >
-            <h3>📚 Referenced Context</h3>
-            <span className="context-logs-toggle">
-              {showContextLogs ? '▼' : '▶'}
-            </span>
-            <span className="context-logs-summary">
-              Auto-injected: {contextLogs.autoInjected.length} |
-              Files: {contextLogs.readFiles.length} |
-              Searches: {contextLogs.globSearches.length + contextLogs.grepSearches.length}
-            </span>
-          </div>
-
-          {showContextLogs && (
-            <div className="context-logs-content">
-              {/* Built-in skills (system) */}
-              <div className="context-log-group">
-                <div className="context-log-title">⚙️ Built-in Skills (System)</div>
-                <ul>
-                  <li className="context-log-item builtin-skill">
-                    keybindings-help
-                  </li>
-                </ul>
-              </div>
-
-              {/* Auto-injected files */}
-              {contextLogs.autoInjected.length > 0 && (
-                <div className="context-log-group">
-                  <div className="context-log-title">🔧 Auto-injected (User Config)</div>
-                  <ul>
-                    {contextLogs.autoInjected.map((file, idx) => (
-                      <li key={idx} className="context-log-item auto-injected">
-                        {file.replace(/^\/Users\/[^/]+/, '~')}
-                      </li>
-                    ))}
-                  </ul>
-                </div>
-              )}
-
-              {/* Read files */}
-              {contextLogs.readFiles.length > 0 && (
-                <div className="context-log-group">
-                  <div className="context-log-title">📖 Read Files</div>
-                  <ul>
-                    {contextLogs.readFiles.map((file, idx) => (
-                      <li key={idx} className="context-log-item read-file">
-                        {file.replace(/^\/Users\/[^/]+/, '~')}
-                      </li>
-                    ))}
-                  </ul>
-                </div>
-              )}
-
-              {/* Glob search */}
-              {contextLogs.globSearches.length > 0 && (
-                <div className="context-log-group">
-                  <div className="context-log-title">🔍 File Search (Glob)</div>
-                  <ul>
-                    {contextLogs.globSearches.map((search, idx) => (
-                      <li key={idx} className="context-log-item glob-search">
-                        <span className="search-pattern">{search.pattern}</span>
-                        <span className="search-path">in {search.searchPath.replace(/^\/Users\/[^/]+/, '~')}</span>
-                      </li>
-                    ))}
-                  </ul>
-                </div>
-              )}
-
-              {/* Grep search */}
-              {contextLogs.grepSearches.length > 0 && (
-                <div className="context-log-group">
-                  <div className="context-log-title">🔎 Content Search (Grep)</div>
-                  <ul>
-                    {contextLogs.grepSearches.map((search, idx) => (
-                      <li key={idx} className="context-log-item grep-search">
-                        <span className="search-pattern">{search.pattern}</span>
-                        <span className="search-path">in {search.searchPath.replace(/^\/Users\/[^/]+/, '~')}</span>
-                      </li>
-                    ))}
-                  </ul>
-                </div>
-              )}
-
-              {contextLogs.sessionId && (
-                <div className="context-log-session">
-                  Session: {contextLogs.sessionId.slice(0, 8)}...
-                </div>
-              )}
-            </div>
-          )}
-        </div>
+        <ContextLogsPanel
+          contextLogs={contextLogs}
+          showContextLogs={showContextLogs}
+          onToggle={handleToggleContextLogs}
+        />
       )}
     </div>
   );
