+++ src/components/tokenTreemap/constants.ts
@@ -1,0 +1,95 @@
+import type { CacheUsageItem } from '../../types';
+
+export type TreemapNode = {
+  name: string;
+  originalName?: string;
+  size?: number;
+  tokens?: number;
+  percentage?: number;
+  color?: string;
+  children?: TreemapNode[];
+  cacheDetails?: CacheUsageItem[];
+  claudeMdPreview?: string;
+  status?: 'warning' | 'good' | 'info' | 'neutral';
+  statusBadge?: string;
+};
+
+export const COLORS = {
+  claudeMd: '#e74c3c',
+  cacheCreation: '#f39c12',
+  cacheRead: '#27ae60',
+  userInput: '#3498db',
+  output: '#9b59b6',
+  section: '#1abc9c',
+} as const;
+
+export const CATEGORY_INFO: Record<string, {
+  friendlyName: string;
+  icon: string;
+  description: string;
+  whatIsIt: string;
+  status: 'warning' | 'good' | 'info' | 'neutral';
+}> = {
+  'CLAUDE.md': {
+    friendlyName: '\u{1f4c4} Config File',
+    icon: '\u{1f4c4}',
+    description: 'AI instructions sent with every request',
+    whatIsIt: 'CLAUDE.md is a set of instructions for the AI. It includes code style, rules, and conventions, and the entire content is sent with every request, incurring costs.',
+    status: 'warning',
+  },
+  'Cache Creation': {
+    friendlyName: '\u{1f195} New Cache',
+    icon: '\u{1f195}',
+    description: 'Cache to be reused in subsequent requests',
+    whatIsIt: 'New context is being stored as cache. It costs now, but will be reused at a 90% discount from the next request onward.',
+    status: 'info',
+  },
+  'Cache Read': {
+    friendlyName: '\u267b\ufe0f Cached',
+    icon: '\u267b\ufe0f',
+    description: '90% cost savings!',
+    whatIsIt: 'Content from previous conversations is being reused. Claude recognizes previously processed content and handles it quickly at a 90% discount.',
+    status: 'good',
+  },
+  'User Input': {
+    friendlyName: '\u{1f4ac} My Input',
+    icon: '\u{1f4ac}',
+    description: 'Your actual input',
+    whatIsIt: 'Your actual prompts and conversation content. This is hard to reduce, but being concise and focused helps efficiency.',
+    status: 'neutral',
+  },
+  'AI Output': {
+    friendlyName: '\u{1f916} AI Response',
+    icon: '\u{1f916}',
+    description: 'AI-generated response',
+    whatIsIt: 'Tokens generated by the AI. Output tokens cost 5x more than input, so request only what you need.',
+    status: 'neutral',
+  },
+};
+
+export const MODEL_PRICING = {
+  'claude-sonnet-4-20250514': { name: 'Claude Sonnet 4', input: 3.0, output: 15.0, cacheRead: 0.3 },
+  'claude-3-5-sonnet-20241022': { name: 'Claude 3.5 Sonnet', input: 3.0, output: 15.0, cacheRead: 0.3 },
+  'claude-3-opus-20240229': { name: 'Claude 3 Opus', input: 15.0, output: 75.0, cacheRead: 1.5 },
+  'claude-3-haiku-20240307': { name: 'Claude 3 Haiku', input: 0.25, output: 1.25, cacheRead: 0.03 },
+  'claude-opus-4-20250514': { name: 'Claude Opus 4', input: 15.0, output: 75.0, cacheRead: 1.5 },
+} as const;
+
+export type ModelId = keyof typeof MODEL_PRICING;
+
+export type TreemapCellProps = {
+  x: number;
+  y: number;
+  width: number;
+  height: number;
+  name?: string;
+  tokens?: number;
+  percentage?: number;
+  color?: string;
+  depth?: number;
+};
+
+export type TreemapTooltipProps = {
+  active?: boolean;
+  payload?: Array<{ payload: TreemapNode }>;
+};
